[REF] mail: introduce chatter and message_list service
This reduces the complexity of thread_service,
and encapsulate self-contained code of chatter-specific
code of thread and message list loading.
