+++ addons/mail/static/src/core/message_service.js
@@ -148,9 +148,11 @@ export class MessageService {
             await this.setDone(message);
         }
         const thread = message.originThread;
-        await this.env.services["mail.thread"].removeFollower(thread.followerOfSelf);
+        await this.env.services["discuss.chatter"].removeFollower(thread.followerOfSelf);
         this.env.services.notification.add(
-            sprintf(_t('You are no longer following "%(thread_name)s".'), { thread_name: thread.name }),
+            sprintf(_t('You are no longer following "%(thread_name)s".'), {
+                thread_name: thread.name,
+            }),
             { type: "success" }
         );
     }
@@ -300,7 +302,7 @@ export class MessageService {
             message.originThread.modelName = data.res_model_name;
         }
         if ("user_follower_id" in data && data.user_follower_id && this.store.self) {
-            this.env.services["mail.thread"].insertFollower({
+            this.env.services["discuss.chatter"].insertFollower({
                 followedThread: message.originThread,
                 id: data.user_follower_id,
                 isActive: true,
