+++ addons/mail/static/src/web/activity/activity_list_popover_item.js
@@ -37,7 +37,7 @@ export class ActivityListPopoverItem extends Component {
         this.state = useState({ hasMarkDoneView: false });
         if (this.props.activity.activity_category === "upload_file") {
             this.attachmentUploader = useAttachmentUploader(
-                this.env.services["mail.thread"].getThread(
+                this.env.services["discuss.chatter"].getThread(
                     this.props.activity.res_model,
                     this.props.activity.res_id
                 )
